testcanvas and rootwidget moved
